Validator metoder tilføjet

+++ src/main/java/org/example/proflow/controller/TaskController.java
@@ -1,7 +1,12 @@
 package org.example.proflow.controller;
 
+import jakarta.servlet.http.HttpSession;
+import org.example.proflow.model.Profile;
+import org.example.proflow.model.Project;
 import org.example.proflow.model.Task;
+import org.example.proflow.service.ProjectService;
 import org.example.proflow.service.TaskService;
+import org.example.proflow.util.Validator;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -11,22 +16,33 @@ import java.util.List;
 
 //TODO TaskController: Rette endpoints og html sider
 //TODO TaskController: Rette exceptions til taskException
-
 @Controller
 @RequestMapping("dashboard/{profileId}/{projectId}")
 public class TaskController {
     //***ATTRIBUTES***--------------------------------------------------------------------------------------------------
     private final TaskService taskService;
+    private final ProjectService projectService;
     private Model model;
 
     //***CONSTRUCTOR***-------------------------------------------------------------------------------------------------
-    public TaskController(TaskService taskService) {
+    public TaskController(TaskService taskService, ProjectService projectService) {
         this.taskService = taskService;
+        this.projectService = projectService;
     }
 
     //***CREATE TASK METHODS***-----------------------------------------------------------------------------------------
     @GetMapping("/{subProjectId}/add-task") //TODO lave tjek med isValid og IsProjectOwned
-    public String addTask(@PathVariable("subProjectId") int subProjectId, Model model) {
+    public String addTask(@PathVariable("subProjectId") int subProjectId, Model model, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");
+        if (!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId);
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())) {
+            return "redirect:/dashboard";
+        }
         model.addAttribute("subProjectId", subProjectId);
         model.addAttribute("task", new Task());
         return "add_task";
@@ -42,15 +58,36 @@ public class TaskController {
 
     //***READ TASK METHODS***-------------------------------------------------------------------------------------------
     @GetMapping("tasks") //TODO skal denne egentlig være en getAllTasksFromSubproject/getTasksFromSubProjectId og hvilken controller klasse skal den så ligge i?
-    //TODO lave tjek med isValid og IsProjectOwned
-    public String getAllTasks(Model model) {
+    public String getAllTasks(Model model, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");
+        if (!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId);
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())) {
+            return "redirect:/dashboard";
+        }
+
         List<Task> tasks = taskService.getAllTasks();
         model.addAttribute("tasks", tasks);
         return "homepage";
     }
 
-    @GetMapping("/task/{taskId}") //TODO lave tjek med isValid og IsProjectOwned
-    public String getTaskById(@PathVariable("taskId") int taskId, Model model) throws SQLException {
+    @GetMapping("/task/{taskId}")
+    public String getTaskById(@PathVariable("taskId") int taskId, Model model, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");
+        if (!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId);
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())) {
+            return "redirect:/dashboard";
+        }
+
         Task task = taskService.getTaskById(taskId);
         model.addAttribute("task", task);
         model.addAttribute("name", task.getName());
@@ -58,8 +95,19 @@ public class TaskController {
     }
 
     //***UPDATE TASK METHODS***-----------------------------------------------------------------------------------------
-    @GetMapping("/task/edit/{taskId}") //TODO lave tjek med isValid og IsProjectOwned
-    public String editTask(@PathVariable("taskId") int taskId, Model model) throws SQLException {
+    @GetMapping("/task/edit/{taskId}")
+    public String editTask(@PathVariable("taskId") int taskId, Model model, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");
+        if (!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId);
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())) {
+            return "redirect:/dashboard";
+        }
+
         Task task = taskService.getTaskById(taskId);
         model.addAttribute("task", task);
 //        model.addAttribute("taskId", task.getId()); //TODO slet de her? 
@@ -77,7 +125,7 @@ public class TaskController {
         return "edit_task";
     }
 
-    @PostMapping("/task/update") //TODO lave tjek med isValid og IsProjectOwned (men skal det gøres på både edit og update?)
+    @PostMapping("/task/update")
     public String updateTask(@ModelAttribute Task task) throws SQLException {
         int taskId = task.getId();
         taskService.updateTask(task);
@@ -85,13 +133,22 @@ public class TaskController {
     }
 
     //***DELETE TASK METHODS***-----------------------------------------------------------------------------------------
-    @PostMapping("/task/delete/{taskId}") //TODO lave tjek med isValid og IsProjectOwned
-    public String deleteTask(@PathVariable("taskId") int taskId) throws SQLException {
+    @PostMapping("/task/delete/{taskId}")  //TODO skal validator være her?
+    public String deleteTask(@PathVariable("taskId") int taskId, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");
+        if (!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId);
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())) {
+            return "redirect:/dashboard";
+        }
         Task task = taskService.getTaskById(taskId);
         taskService.deleteTask(taskId);
         return "redirect:/subproject";
     }
-
 
     //***END***---------------------------------------------------------------------------------------------------------
 }
